Added a logging library spdlog. Can specify a log level when launching webx  (either debug or trace. Otherwise defaults to info). Added spdlog to Cmake. Replaced the majority of printf method calls to logger method calls. Added a little testMouse program for testing cursor movement and mouse clicking

+++ src/display/WebXWindow.h
@@ -73,12 +73,12 @@ public:
     void enableDamage();
     void disableDamage();
 
-    const uint64_t getWindowChecksum() const {
+    const uint32_t getWindowChecksum() const {
         return this->_windowChecksum;
     }
 
 private:
-    uint64_t calculateImageChecksum(XImage * image);
+    uint32_t calculateImageChecksum(XImage * image);
 
 private:
     Display * _display;
